LayoutTests/platform/ios-simulator-wk2/fast/events/touch/gesture/long-press-on-draggable-element-triggers-drag-expected.txt

CONSOLE MESSAGE: line 18: TypeError: undefined is not a function (evaluating 'internals.settings.setTouchDragDropEnabled(true)')
This test checks that a long press gesture on a draggable element starts a drag/drop session.

Testing
FAIL
Drag me
